update: extracted ThemeToggle component to separate file
